Self insert to support a lot of utf chars
WARNING: This commit breaks backwards compatibility.
All keysmaps now uses strings as keys like "a", "b", "pgdn", etc,
instead of keywords like :a, :b, :pagedown.

A new special keymap keyword has been introduced :selfinsert
The value of this should be a function taking a string as
input. When the keymapping is resolved and there are no dedicated
entry for the given input, the selfinsert function will be called.
This allows any character input to be used directly in the text.

Hopefully a lot of asian characters will be supported this way.
At least in the tty implementation.