Release of 2.0
I am delighted to announce the release of Liquid 2.0.

It is a rewrite from scratch, based on the experience from Liquid 1. This also means that it has breaking changes. Liquid 2 can be put in place of Liquid 1 with some adjustments. If you run into any problems in that regard, please contact me directly and I will try to help.

So, what is new:

 * How it works internally is more consistant and easier to interpret, which will help contributers on the long run.
 * Functionality is loaded into the editor in a more consistent way, making extension creation easier and more "live".
 * The namespaces have been shortened.
 * Primary inspiration is Vim. The aim is to make it possible to switch between Vim and Liquid without hazzle.
 * Emacs and Spacemacs ideas have also been implemented where they do not conflict.
 * Of cause Liquid can be embedded in your project, but it is also possible to use a "jack-in" like connection to a socket REPL.
 * Endpoints for integration with parinfer-rust.

Current state:

The basic functionality works. Please have the Cheatsheet nearby for support.

Of cause not every command and keystroke from Vim, Emacs and Spacemacs are in place, but it should be relatively easy to add more along the way.

It has been made ready to run in a NodeJS environment as well. This part is not mature yet.

Whats next:

Improve piece by piece as needed and focus on extensions.