Eval buffer with c p f and improved keybinding resolve
To evaluate the content of the current buffer type c p f. It evaluates
the content of the buffer, not the file, so saving is not needed beforehand.

c p F have been implemented to evaluate the current file raw.

From the list of major modes the keybindings are now merged when resolved.
This allows better handling of nested keybindings.